# Reduce pandas dataframe overhead.