Related: https://github.com/dmlc/xgboost/issues/10882

The performance is not back to 1.7 yet, but the regression is not so severe. It's still far away from numpy performance. The result is obtained using the script in the referenced issue.
- Lazy import some of the pandas modules.
- Merge type check into type conversion.
- Early return of nullable type check.

## 1.7
```
DataFrame iter_time:  0.2178664207458496 ms
DataFrame iter_time:  0.21868157386779785 ms
DataFrame iter_time:  0.23699164390563965 ms
```

## PR
```
DataFrame iter_time:  0.27150797843933105 ms
DataFrame iter_time:  0.28162598609924316 ms
DataFrame iter_time:  0.278597354888916 ms
```

## Master
```
DataFrame iter_time:  0.8856163024902344 ms
DataFrame iter_time:  0.9370462894439697 ms
DataFrame iter_time:  0.9501245021820068 ms
```